+++ chip8_emulator/src/lib/OpCodes/OpCode6XNN.java
@@ -9,7 +9,7 @@ import lib.MachineState;
 
 public class OpCode6XNN extends OpCodeVXNN {
     @Override
-    public void execute(MachineState mc, int myCode, int vX, int NN) {
+    public void execute(MachineState mc, int myCode, int vX, byte NN) {
         mc.V[vX] = NN;
     }
 }
