+++ chip8_emulator/src/lib/OpCodes/OpCode7XNN.java
@@ -9,9 +9,9 @@ import lib.MachineState;
 
 public class OpCode7XNN extends OpCodeVXNN {
     @Override
-    public void execute(MachineState mc, int myCode, int vX, int NN) {
+    public void execute(MachineState mc, int myCode, int vX, byte NN) {
         mc.V[vX] += NN;
-        mc.V[vX] = mc.V[vX] % 0xFF;
+        mc.V[vX] = (byte) (mc.V[vX] % 0xFF);
 
     }
 }
